// types/author.ts
interface Author {
	name: string;
	avatar: string;
	github: string;
	heroku: string;
	vercel: string;
	props: string | number;
	antdesign: string;
	apollo: string;
  auth0: string;
  graphql: string;
  next: string;
  react: string;
  tailwindcss: string;
	typescript: string;
	wordpress: string;
}

export default Author;